LayoutTests/imported/w3c/web-platform-tests/web-animations/responsive/textIndent-expected.txt


FAIL textIndent responsive to style changes assert_not_equals: got disallowed value "100px hanging"
FAIL textIndent responsive to inherited textIndent changes assert_equals: expected "200px hanging each-line" but got "150px hanging each-line"
